requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.3.0
certifi==2026.2.25
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.5
daphne==4.1.0
Django==4.2
django-cors-headers==4.9.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.0
requests==2.32.5
service-identity==24.2.0
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.2
